requirements.txt
numpy>=1.26.0
pandas>=2.1.0
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1
transformers>=4.36.0
torch>=2.1.0
Pillow>=10.0.0
Flask>=2.3.2
vaderSentiment>=3.3.2
requests>=2.31.0
setuptools>=68.0.0
wheel>=0.41.0 